requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
sqlalchemy>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pyyaml>=6.0
pydantic
python-dotenv
最关键的解析器，供 dialect_translator 转化为 AST

mindsdb-sql-parser>=0.1.0